+++ tests/plugins/undo/integrations.js
@@ -21,9 +21,7 @@
 				this.undoManager = this.editor.undoManager;
 
 				// For each TC we want to reset undoManager.
-				this.undoManager.reset();
-				// Force to reset inputFired counter, as some TCs may produce leftovers.
-				this.undoManager.editingHandler.resetCounter();
+				this.editor.resetUndo();
 			},
 
 			// (#12327)